requirements.txt
# Core ML/DL dependencies
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.8.0
keras>=2.8.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.1.0
scikit-image>=0.19.0

# Data handling and visualization
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.3.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
argparse
json5
pyyaml>=6.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Model evaluation
segmentation-models-pytorch>=0.2.0
monai>=0.9.0

# Optional GPU acceleration
# Uncomment if using CUDA
# torch-audio
# torch-text